refactor: constant 폴더 분리
