+++ src/utils/constant/folder.ts
@@ -1,0 +1,6 @@
+const DEFAULT_FOLDER = {
+  name: '전체',
+  id: 'all',
+};
+
+export default DEFAULT_FOLDER;
